Add kpm push helm charts

+++ kpm/commands/push.py
@@ -1,6 +1,8 @@
+import argparse
 import os
 import json
 import base64
+import kpm.utils
 import kpm.registry
 from kpm.manifest_jsonnet import ManifestJsonnet
 from kpm.packager import pack_kub
@@ -17,39 +19,53 @@ class PushCmd(CommandBase):
         self.force = options.force
         self.manifest = None
         self.media_type = options.media_type
-
+        self.version = options.version
+        self.package_name = options.name
+        self.filter_files = True
         super(PushCmd, self).__init__(options)
 
     @classmethod
     def _add_arguments(cls, parser):
         cls._add_registryhost_option(parser)
         cls._add_mediatype_option(parser)
-        cls._add_packagename_option(parser)
         cls._add_packageversion_option(parser)
+        parser.add_argument('-n', "--name", default=None, action=kpm.command.PackageName, help="package-name")
         parser.add_argument("-f", "--force", action='store_true', default=False,
                             help="force push")
 
     def _call(self):
         r = kpm.registry.Registry(self.registry_host)
-        # @TODO: Override organization
-        self.manifest = ManifestJsonnet()
+        if self.media_type in ["kpm", "kpm-compose"]:
+            self.manifest = ManifestJsonnet()
+            if not self.package_name:
+                self.package_name = self.manifest.package['name']
+            if not self.version or self.version == "default":
+                self.version = self.manifest.package['version']
+        else:
+            self.filter_files = False
+
+        if self.version is None or self.version == "default":
+            raise argparse.ArgumentTypeError("Missing option: --version")
+        if self.package_name is None:
+            raise argparse.ArgumentTypeError("Missing option: --name")
+
+        filename = kpm.utils.package_filename(self.package_name, self.version, self.media_type)
         # @TODO: Pack in memory
-        kubepath = os.path.join(".", self.manifest.package_name() + "kub.tar.gz")
+        kubepath = os.path.join(".", filename + ".tar.gz")
         pack_kub(kubepath)
         f = open(kubepath, 'rb')
-        body = {"name": self.manifest.package['name'],
-                "release": self.manifest.package['version'],
+        body = {"name": self.package_name,
+                "release": self.version,
                 "media_type": self.media_type,
                 "blob": base64.b64encode(f.read())}
-        r.push(self.manifest.package['name'], body, self.force)
+        r.push(self.package_name, body, self.force)
         f.close()
         os.remove(kubepath)
 
     def _render_json(self):
-        print json.dumps({"package": self.manifest.package['name'],
-                          "version": self.manifest.package['version'],
+        print json.dumps({"package": self.package_name,
+                          "version": self.version,
                           "media_type": self.media_type})
 
     def _render_console(self):
-        print "package: %s (%s) pushed" % (self.manifest.package['name'],
-                                           self.manifest.package['version'])
+        print "package: %s (%s) pushed" % (self.package_name, self.version)
